requirements.txt
streamlit
langchain
langchain-groq
langchain-openai
langchain-community
langchain-core
python-dotenv
faiss-cpu
PyPDF2
openai
langchain-huggingface
pypdf